cloud/inference_deployment/requirements.txt
numpy
torch
pillow
transformers